Moved Signup test files to own directory
