+++ frontend/tests/pages/Signup/SignupPage.integration.test.jsx
@@ -2,10 +2,10 @@ import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, vi } from "vitest";
 import { MemoryRouter, useNavigate, Route, Routes } from "react-router-dom";
-import * as authentication from "../../src/services/authentication";
-import { SignupPage } from "../../src/pages/Signup/SignupPage";
-import { LoginPage } from "../../src/pages/Login/LoginPage";
-import HomePage from "../../src/pages/Home/HomePage";
+import * as authentication from "../../../src/services/authentication";
+import { SignupPage } from "../../../src/pages/Signup/SignupPage";
+import { LoginPage } from "../../../src/pages/Login/LoginPage";
+import HomePage from "../../../src/pages/Home/HomePage";
 
 const signupSpy = vi.spyOn(authentication, "signup");
